requirements.txt


streamlit==1.25.0
protobuf==3.20.3
tensorflow==2.13.0
numpy==1.24.3
Pillow==9.5.0
scipy==1.10.1
